土地容积率调整

+++ assess-app/src/main/java/com/copower/pmcc/assess/dal/basis/mapper/BasicEstateLandCategoryInfoMapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface BasicEstateLandCategoryInfoMapper {
-    int countByExample(BasicEstateLandCategoryInfoExample example);
+    long countByExample(BasicEstateLandCategoryInfoExample example);
 
     int deleteByExample(BasicEstateLandCategoryInfoExample example);
